addBook btn listener rearrange
